chore: Replace Milkdown with Quill
I found Milkdown a bit abscure and dificult to adapt.
Decided to start from scratch with something more bare-bones.
Chose Quill.  The hope is I have a better understanding of how the
Editor is funtioning, so I can better adapt it to what i need.

- Had to fork quill-delta-to-markdown to add support for code blocks.
Not sure is package is still maintained, but PR is open;
https://github.com/frysztak/quill-delta-to-markdown/pull/3
- extracted LlmPrompting.vue from EditorToolbar.vue

removed all ref to Milkdown
