+++ package.json
@@ -12,22 +12,13 @@
     "docs-gen": "cd src-tauri && cargo doc --no-deps && cp -r target/doc/* ../docs/ && open ../docs/nova_notes/index.html"
   },
   "dependencies": {
-    "@milkdown/core": "^7.3.5",
-    "@milkdown/ctx": "^7.3.5",
-    "@milkdown/plugin-clipboard": "^7.3.6",
-    "@milkdown/plugin-listener": "^7.3.5",
-    "@milkdown/plugin-slash": "^7.3.6",
-    "@milkdown/plugin-tooltip": "^7.3.6",
-    "@milkdown/plugin-trailing": "^7.3.6",
-    "@milkdown/preset-commonmark": "^7.3.5",
-    "@milkdown/prose": "^7.3.5",
-    "@milkdown/theme-nord": "^7.3.5",
-    "@milkdown/transformer": "^7.3.5",
-    "@milkdown/vue": "^7.3.5",
-    "@prosemirror-adapter/vue": "^0.2.2",
     "@tauri-apps/api": "^1",
     "lodash-es": "^4.17.21",
-    "marked": "^12.0.1",
+    "marked": "^12.0.2",
+    "md-to-quill-delta": "^1.1.1",
+    "quill": "^2.0.2",
+    "quill-delta-to-markdown": "git+https://github.com/samkeen/quill-delta-to-markdown.git#master",
+    "quilljs-markdown": "^1.2.0",
     "tauri-plugin-log-api": "github:tauri-apps/tauri-plugin-log#v1",
     "tauri-plugin-store-api": "github:tauri-apps/tauri-plugin-store#v1",
     "vue": "^3.3.4",
