+++ src/resources/default-md.js
@@ -1,0 +1,12 @@
+export const defaultMarkdownText = `# sample markdown text 
+
+## Mark Features 
+- Minimal and Quick
+- Supports commonmark standards
+- built to be to the point and add no overhead for people who simple enjoy writing in markdown 
+
+## Motivation 
+I write a lot, which you can read on [reaper.is](https://reaper.is) and this tool was primary 
+built to act as a quick markdown app that I could access from anywhere.
+
+`;
