+++ src/views/Editor.vue
@@ -3,110 +3,101 @@
     <div class="flex flex-col flex-1 overflow-hidden">
       <EditorToolbar/>
       <div class="flex-1 overflow-y-auto p-4">
-        <template v-if="noteLoaded">
-          <div class="root">
-            <MilkdownProvider>
-              <ProsemirrorAdapterProvider>
-                <MilkdownEditor :initialValue="noteText" :readonly="false" @update="handleNoteUpdate"/>
-              </ProsemirrorAdapterProvider>
-            </MilkdownProvider>
-          </div>
-        </template>
-        <template v-else>
-          <div class="text-center">Loading note...</div>
-        </template>
+        <div class="root">
+          <Editor
+              v-on:change="debouncedHandleChange"
+              v-bind:initialDocContent="docState.noteObj ? docState.noteObj.text : ''"
+          ></Editor>
+        </div>
       </div>
     </div>
     <Sidebar :note-id="noteId"/>
   </div>
 </template>
-<!--<script setup>-->
-<!--import Milkdown from './components/MilkdownEditor.vue'-->
-<!--import {MilkdownProvider} from '@milkdown/vue';-->
-<!--import {ProsemirrorAdapterProvider} from '@prosemirror-adapter/vue';-->
-<!--</script>-->
-
-<!--<template>-->
-<!--  <MilkdownProvider>-->
-<!--    <ProsemirrorAdapterProvider>-->
-<!--      <Milkdown/>-->
-<!--    </ProsemirrorAdapterProvider>-->
-<!--  </MilkdownProvider>-->
-<!--</template>-->
 
 <script setup>
 /**
- * For Layout, see: https://claude.ai/chat/72a74a4e-b343-49f7-a76e-0eabaeb0c1d7
- *
- * ## Autosave Functionality
- * 1. When the user types or modifies the content in the Milkdown editor, the update event is emitted, triggering the handleNoteUpdate function.
- * 2. Inside the handleNoteUpdate function, the noteText value is updated with the new content received from the event.
- * 3. The debouncedSaveNote function is called, which is a debounced version of the save_note function.
- * 4. The debounce function from lodash-es is used to create the debouncedSaveNote function. It takes two arguments: the function to be debounced (save_note) and the delay in milliseconds (autosaveDelay).
- * 5. The debounce function works as follows:
- *    - When debouncedSaveNote is called, it starts a timer with the specified delay (1000ms in this case).
- *    - If debouncedSaveNote is called again within the delay period, the previous timer is canceled, and a new timer is started.
- *    - If no further calls to debouncedSaveNote occur within the delay period, the save_note function is finally invoked.
- * 6. This means that the save_note function is not called immediately every time the user types. Instead, it waits for a pause in the user's typing activity. If the user continues typing within the specified delay (1000ms), the previous autosave is canceled, and a new autosave is scheduled.
- * 7. Once the user stops typing and the delay period passes without any further modifications, the save_note function is invoked, saving the note.
+ * Handles the application state, including the Markdown code and state management.
+ * It also listens for the change event emitted by the Editor component to update
+ * the code in the state and save it to local storage.
+ * [ ] TODO Utilize keydown event listener
  */
-import Sidebar from '../components/Sidebar.vue'
+import Editor from "../components/QuillEditor.vue";
 import EditorToolbar from "../components/EditorToolbar.vue";
-import {useRoute} from 'vue-router';
+import Sidebar from "../components/Sidebar.vue";
+// Import the necessary composition API functions from Vue
+import {onMounted, onUnmounted, reactive, ref} from "vue";
+import {useRoute} from "vue-router";
+import {debug, info} from "tauri-plugin-log-api";
 import {invoke} from "@tauri-apps/api/tauri";
-import {onMounted, ref} from "vue";
-import {debounce} from 'lodash-es';
-import {upsertNote} from '../lib/notebook.js';
-import {info} from "tauri-plugin-log-api";
-import {MilkdownProvider} from "@milkdown/vue";
-import MilkdownEditor from "../components/MilkdownEditor.vue";
-import {ProsemirrorAdapterProvider} from '@prosemirror-adapter/vue';
-
-let noteText = ref('');
-let originalNoteText = ref('');
-let noteLoaded = ref(false);
-
-const autosaveDelay = 1000; // Adjust the delay as needed (in milliseconds)
+import {upsertNote} from "../lib/notebook.js";
+import {debounce} from "lodash-es";
 
 const route = useRoute();
 let noteId = ref(route.params.id || null);
 
-const debouncedSaveNote = debounce(async () => {
-  await handleSaveNote();
-}, autosaveDelay);
-
-function handleNoteUpdate(event) {
-  noteText.value = event;
-  debouncedSaveNote();
-}
-
-async function handleSaveNote() {
-  if (noteText.value !== originalNoteText.value) {
-    const savedNoteId = await upsertNote(noteId.value, noteText.value);
-    if (!noteId.value && savedNoteId) {
-      noteId.value = savedNoteId;
-    }
-  } else {
-    info("Note content has not changed. Skipping save.");
-  }
-}
-
-onMounted(async () => {
+// Function to get the default code
+const loadDocContent = async (noteId) => {
+  let noteObj = null;
   if (noteId.value) {
     info(`The note id: ${noteId.value}`);
     try {
-      const note = await invoke("get_note_by_id", {id: noteId.value});
-      noteText.value = note.text;
-      originalNoteText.value = note.text;
-      noteLoaded.value = true;
+      noteObj = await invoke("get_note_by_id", {id: noteId.value});
+      console.log("Note loaded:", noteObj);
     } catch (error) {
       console.error("Failed getting note:", error);
-      // Handle the error as needed, e.g., show a user-friendly message
     }
   } else {
     info("The note id is not defined");
-    noteText.value = ''; // Initialize noteText.value with an empty string for new notes
-    noteLoaded.value = true;
   }
+  return noteObj;
+};
+
+const docState = reactive({
+  noteId: noteId,
+  copied: false,
+  /**
+   * id: String,
+   * text: String,
+   * categories: HashSet<Category>,
+   * created: i64,
+   * modified: i64,
+   */
+  noteObj: null, // Initialize noteObj as null
 });
+
+// Lifecycle hook: Mounted
+onMounted(async () => {
+  // Load the note content asynchronously
+  const noteObj = await loadDocContent(noteId);
+  docState.noteObj = noteObj;
+  // Add a keydown event listener to the document
+  document.addEventListener("keydown", shortcutListener.bind(this));
+});
+
+// Lifecycle hook: Unmounted
+onUnmounted(() => {
+  // Remove the keydown event listener from the document
+  document.removeEventListener("keydown", shortcutListener);
+});
+
+// Event handler for keydown events
+function shortcutListener(e) {
+  console.log("shortcut listener", e);
+}
+
+// Create a debounced version of the handleChange function
+const debouncedHandleChange = debounce(handleChange, 1500);
+
+// Event handler for the 'change' event emitted by the Editor component
+async function handleChange(docContent) {
+  // docState.noteObj.text = docContent;
+  debug(`HANDLED CHANGE: ${docContent}`);
+  info(`Saving note with id: ${noteId.value}`);
+  const savedNoteId = await upsertNote(noteId.value, docContent);
+  // If new note, populate the id
+  if (!noteId.value && savedNoteId) {
+    noteId.value = savedNoteId;
+  }
+}
 </script>
